+++ lib/resizable_splitter.dart
@@ -11,6 +11,7 @@ library;
 
 export 'src/resizable_splitter.dart';
 export 'src/resizable_splitter_theme.dart';
+export 'src/split_animation.dart';
 export 'src/split_divider_style.dart';
 export 'src/split_layout.dart';
 export 'src/split_pane_constraints.dart';